feat(dng): crop into active area; simplified demosaicing
Many TIFF/DNG options become simplified by cropping into the active area
prior to processing. Demosaicing is moved into a separate file & no
longer operates on planes that have to get re-interleaved.
